+++ shop/shipping/backends/flat_rate.py
@@ -5,8 +5,9 @@ from django.conf import settings
 from django.conf.urls.defaults import patterns, url
 from django.shortcuts import render_to_response
 from django.template import RequestContext
+from django.utils.translation import ugettext_lazy as _
 
-from shop.util.decorators import on_method, shop_login_required
+from shop.util.decorators import on_method, shop_login_required, order_required
 
 
 class FlatRateShipping(object):
@@ -16,6 +17,7 @@ class FlatRateShipping(object):
     """
     url_namespace = 'flat'
     backend_name = 'Flat rate'
+    backend_verbose_name = _('Flat rate')
 
     def __init__(self, shop):
         self.shop = shop  # This is the shop reference, it allows this backend
@@ -23,6 +25,7 @@ class FlatRateShipping(object):
         self.rate = getattr(settings, 'SHOP_SHIPPING_FLAT_RATE', '10')
 
     @on_method(shop_login_required)
+    @on_method(order_required)
     def view_process_order(self, request):
         """
         A simple (not class-based) view to process an order.