+++ shop/util/decorators.py
@@ -1,8 +1,13 @@
 """Decorators for the django-shop application."""
+from functools import wraps
+
 from django.contrib.auth import REDIRECT_FIELD_NAME
 from django.contrib.auth.decorators import user_passes_test
+from django.http import HttpResponseRedirect
 
 from shop.util.login_mixin import get_test_func
+from shop.util.order import get_order_from_request
+from shop.models.ordermodel import Order
 
 
 def on_method(function_decorator):
@@ -37,3 +42,34 @@ def shop_login_required(function=None, redirect_field_name=REDIRECT_FIELD_NAME,
     if function:
         return actual_decorator(function)
     return actual_decorator
+
+
+def order_required(redirect_url='/'):
+    """
+    Ensures that an non-complete order exists before carrying out any
+    additional functions that rely on one.
+
+    If an order does not exist the browser will be redirected to another page
+    supplied in the optional keyword argument `redirect_url`.
+
+    Usage:
+    @order_required
+    def some_view(...
+
+    OR:
+    @order_required(redirect_url='/some/path/')
+    def some_view(...
+    """
+    if callable(redirect_url):
+        func = redirect_url
+        decorator = order_required()
+        return decorator(func)
+
+    def decorator(func):
+        def inner(request, *args, **kwargs):
+            order = get_order_from_request(request)
+            if order is None or getattr(order, 'status', Order.COMPLETED) == Order.COMPLETED:
+                return HttpResponseRedirect(redirect_url)
+            return func(request, *args, **kwargs)
+        return wraps(func)(inner)
+    return decorator