+++ shop/views/checkout.py
@@ -108,8 +108,7 @@ class CheckoutSelectionView(LoginMixin, ShopTemplateView):
         if not form:
             # Create a dynamic Form class for the model specified as the
             # address model
-            form_class = model_forms.modelform_factory(
-                AddressModel, exclude=['user_shipping', 'user_billing'])
+            form_class = self.get_billing_form_class()
 
             # Try to get a shipping address instance from the request (user or
             # session))
@@ -211,21 +210,14 @@ class ThankYouView(LoginMixin, ShopTemplateView):
     def get_context_data(self, **kwargs):
         ctx = super(ShopTemplateView, self).get_context_data(**kwargs)
 
-        # Set the order status:
+        # put the latest order in the context only if it is completed
         order = get_order_from_request(self.request)
-        if order:
-            order.status = Order.COMPLETED
-            order.save()
-            completed.send(sender=self, order=order)
-        else:
-            order = Order.objects.get_latest_for_user(self.request.user)
-            #TODO: Is this ever the case?
-        ctx.update({'order': order, })
-
-        # Empty the customers basket, to reflect that the purchase was
-        # completed
-        cart_object = get_or_create_cart(self.request)
-        cart_object.empty()
+        if order and order.status == Order.COMPLETED:
+            ctx.update({'order': order, })
+            # Empty the customers basket, to reflect that the purchase was
+            # completed
+            cart_object = get_or_create_cart(self.request)
+            cart_object.empty()
 
         return ctx